# Conflicts:
#	demisto_sdk/commands/common/constants.py
#	demisto_sdk/commands/common/tests/tools_test.py
#	demisto_sdk/commands/common/tools.py
#	demisto_sdk/commands/format/format_module.py
#	demisto_sdk/commands/format/update_generic.py
#	demisto_sdk/commands/format/update_generic_yml.py
#	demisto_sdk/commands/format/update_layout.py
#	demisto_sdk/commands/format/update_playbook.py
#	demisto_sdk/commands/secrets/secrets.py
#	demisto_sdk/tests/integration_tests/format_integration_test.py

+++ demisto_sdk/commands/init/contribution_converter.py
@@ -21,6 +21,12 @@ from demisto_sdk.commands.common.tools import (LOG_COLORS, capital_case,
                                                get_child_files,
                                                get_content_path)
 from demisto_sdk.commands.format.format_module import format_manager
+from demisto_sdk.commands.generate_docs.generate_integration_doc import \
+    generate_integration_doc
+from demisto_sdk.commands.generate_docs.generate_playbook_doc import \
+    generate_playbook_doc
+from demisto_sdk.commands.generate_docs.generate_script_doc import \
+    generate_script_doc
 from demisto_sdk.commands.split_yml.extractor import Extractor
 
 
@@ -42,12 +48,13 @@ class ContributionConverter:
         dir_name (str): The directory name of a pack's containing folder
         create_new (bool): True if creating a new pack (default), False if updating an existing pack
         gh_user (str): The github username of the person contributing the pack
+        readme_files (List[str]): The readme files paths that is generated for new content items.
     """
     DATE_FORMAT = '%Y-%m-%dT%H:%M:%SZ'
 
     def __init__(self, name: str = '', contribution: Union[str] = None, description: str = '', author: str = '',
                  gh_user: str = '', create_new: bool = True, pack_dir_name: Union[str] = None,
-                 base_dir: Union[str] = None):
+                 base_dir: Union[str] = None, no_pipenv: bool = False):
         """Initializes a ContributionConverter instance
 
         Note that when recieving a contribution that is an update to an existing pack that the values of 'name',
@@ -67,6 +74,7 @@ class ContributionConverter:
             base_dir (Union[str], optional): Used to explicitly pass the path to the top-level directory of the
                 local content repo. If no value is passed, the `get_content_path()` function is used to determine
                 the path. Defaults to None.
+
         """
         self.configuration = Configuration()
         self.contribution = contribution
@@ -75,6 +83,7 @@ class ContributionConverter:
         self.gh_user = gh_user
         self.contrib_conversion_errs: List[str] = []
         self.create_new = create_new
+        self.no_pipenv = no_pipenv
         base_dir = base_dir or get_content_path()
         self.packs_dir_path = os.path.join(base_dir, 'Packs')
         if not os.path.isdir(self.packs_dir_path):
@@ -88,6 +97,7 @@ class ContributionConverter:
         self.pack_dir_path = os.path.join(self.packs_dir_path, self.dir_name)
         if not os.path.isdir(self.pack_dir_path):
             os.makedirs(self.pack_dir_path)
+        self.readme_files: List[str] = []
 
     @staticmethod
     def format_pack_dir_name(name: str) -> str:
@@ -221,18 +231,66 @@ class ContributionConverter:
                 for _, _, files in os.walk(src_path, topdown=False):
                     for name in files:
                         src_file_path = os.path.join(src_path, name)
-                        shutil.move(src_file_path, dst_path)
+                        dst_file_path = os.path.join(dst_path, name)
+                        shutil.move(src_file_path, dst_file_path)
+                shutil.rmtree(src_path, ignore_errors=True)
             else:
                 # replace dst folder with src folder
                 shutil.move(src_path, dst_path)
 
     def format_converted_pack(self) -> None:
         """Runs the demisto-sdk's format command on the pack converted from the contribution zipfile"""
-        click.echo(f'Executing \'format\' on the restructured contribution zip files at "{self.pack_dir_path}"')
+        click.echo(
+            f'Executing \'format\' on the restructured contribution zip new/modified files at {self.pack_dir_path}'
+        )
         from_version = '6.0.0' if self.create_new else ''
         format_manager(
-            input=self.pack_dir_path, from_version=from_version, no_validate=True, update_docker=True, assume_yes=True
+            from_version=from_version, no_validate=True, update_docker=True, verbose=True, assume_yes=True
         )
+
+    def generate_readme_for_pack_content_item(self, yml_path: str) -> None:
+        """ Runs the demisto-sdk's generate-docs command on a pack content item
+
+        Args:
+            yml_path: str: Content item yml path.
+        """
+        file_type = find_type(yml_path)
+        file_type = file_type.value if file_type else file_type
+        if file_type == 'integration':
+            generate_integration_doc(yml_path)
+        if file_type == 'script':
+            generate_script_doc(input_path=yml_path, examples=[])
+        if file_type == 'playbook':
+            generate_playbook_doc(yml_path)
+
+        dir_output = os.path.dirname(os.path.realpath(yml_path))
+        readme_path = os.path.join(dir_output, 'README.md')
+        self.readme_files.append(readme_path)
+
+    def generate_readmes_for_new_content_pack(self):
+        """
+        Generate the readme files for a new content pack.
+        """
+        for pack_subdir in get_child_directories(self.pack_dir_path):
+            basename = os.path.basename(pack_subdir)
+            if basename in {SCRIPTS_DIR, INTEGRATIONS_DIR}:
+                directories = get_child_directories(pack_subdir)
+                for directory in directories:
+                    files = get_child_files(directory)
+                    for file in files:
+                        file_name = os.path.basename(file)
+                        if file_name.startswith('integration-') \
+                           or file_name.startswith('script-') \
+                           or file_name.startswith('automation-'):
+                            unified_file = file
+                            self.generate_readme_for_pack_content_item(unified_file)
+                            os.remove(unified_file)
+            elif basename == 'Playbooks':
+                files = get_child_files(pack_subdir)
+                for file in files:
+                    file_name = os.path.basename(file)
+                    if file_name.startswith('playbook') and file_name.endswith('.yml'):
+                        self.generate_readme_for_pack_content_item(file)
 
     def convert_contribution_to_pack(self, files_to_source_mapping: Dict = None):
         """Create or updates a pack in the content repo from the contents of a contribution zipfile
@@ -266,8 +324,12 @@ class ContributionConverter:
                 basename = os.path.basename(pack_subdir)
                 if basename in {SCRIPTS_DIR, INTEGRATIONS_DIR}:
                     self.content_item_to_package_format(
-                        pack_subdir, del_unified=True, source_mapping=files_to_source_mapping
+                        pack_subdir, del_unified=(not self.create_new), source_mapping=files_to_source_mapping
                     )
+
+            if self.create_new:
+                self.generate_readmes_for_new_content_pack()
+
             # format
             self.format_converted_pack()
         except Exception as e:
@@ -326,18 +388,27 @@ class ContributionConverter:
                         os.makedirs(output_dir, exist_ok=True)
                         extractor = Extractor(input=content_item_file_path, file_type=file_type, output=output_dir,
                                               no_readme=True, base_name=base_name,
-                                              no_auto_create_dir=(not autocreate_dir))
+                                              no_auto_create_dir=(not autocreate_dir), no_pipenv=self.no_pipenv)
 
                     else:
-                        extractor = Extractor(input=content_item_file_path,
-                                              file_type=file_type, output=content_item_dir)
+                        extractor = Extractor(input=content_item_file_path, file_type=file_type,
+                                              output=content_item_dir, no_pipenv=self.no_pipenv)
                     extractor.extract_to_package_format()
                 except Exception as e:
                     err_msg = f'Error occurred while trying to split the unified YAML "{content_item_file_path}" ' \
                               f'into its component parts.\nError: "{e}"'
                     self.contrib_conversion_errs.append(err_msg)
-                if del_unified:
-                    os.remove(content_item_file_path)
+                finally:
+                    output_path = extractor.get_output_path()
+                    if self.create_new:
+                        # Moving the unified file to its package.
+                        shutil.move(content_item_file_path, output_path)
+                    if del_unified:
+                        if os.path.exists(content_item_file_path):
+                            os.remove(content_item_file_path)
+                        moved_unified_dst = os.path.join(output_path, child_file_name)
+                        if os.path.exists(moved_unified_dst):
+                            os.remove(moved_unified_dst)
 
     def create_pack_base_files(self):
         """
